refactored single test unit tests to ease phasing in of new noise models
